feat: Add custom hook to detect Shift key state for improved keyboard shortcuts
